WIP: Add download_type.dart and refactor download_info.dart
- Create tests for DownloadTypeExtension and downloadTypeFromMimeType.
- Moved DownloadInfo into models/download_info.dart as it is more of a model versus a value.
